% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunePairMarker-package.R
\docType{package}
\name{immunePairMarker-package}
\alias{immunePairMarker}
\alias{immunePairMarker-package}
\title{immunePairMarker}
\description{
Tools for pairwise gene marker analysis in immune response studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/twl-00/immunePairMarker}
  \item Report bugs at \url{https://github.com/twl-00/immunePairMarker/issues}
}

}
\author{
\strong{Maintainer}: twl-00 \email{tanwenlu2000@163.com}

}
\keyword{internal}
